pyproject.toml
[project]
name = "xhptdc8_py"
version = "0.1.0"
description = "Python bindings for the XHPTDC8 TDC driver"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cython>=3.0.12",
    "numpy>=2.2.2",
    "pyyaml>=6.0.2",
    "setuptools>=75.8.0",
]

[build-system]
requires = ['setuptools>=75.8.0', "cython"]
build-backend = 'setuptools.build_meta'

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
line-length = 88
indent-width = 4
target-version = "py313"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "SIM"]
ignore = []
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = 40
